+++ R/clock_chart_col.R
@@ -9,7 +9,7 @@
 #'
 #' Add
 #' or modify legend by `theme(legend.position = "POSITION")`; the valid
-#' postion names in `ggplot2` are `top`, `bottom`, `right`, and `left`,
+#' position names in `ggplot2` are `top`, `bottom`, `right`, and `left`,
 #' excluding more complex options.
 #'
 #' @seealso