+++ R/clock_chart_qnt.R
@@ -9,7 +9,7 @@
 #' `ggplot2::labs(color = "TITLE 1", size = "TITLE2")`.
 #' Add
 #' or modify legend by `theme(legend.position = "POSITION")`; the valid
-#' postion names in `ggplot2` are `top`, `bottom`, `right`, and `left`,
+#' position names in `ggplot2` are `top`, `bottom`, `right`, and `left`,
 #' excluding more complex options.
 #'
 #' @seealso