+++ R/plan_week.R
@@ -1,4 +1,4 @@
-#' Plan Activities on a Week with Events on Each Day
+#' Visually Plan Activities on a Week with Events on Each Day
 #'
 #' This function plots works corresponding to each day of the week.
 #'