+++ R/year_chart.R
@@ -1,4 +1,5 @@
 #' Plot Values on Each Month of a Year
+#'
 #' This function plots values corresponding to each month on a rose plot.
 #'
 #'