+++ NEWS.md
@@ -1,11 +1,12 @@
 # clockplot (development version)
 
 * Added US Accidents data and example. 
+* Added error for non-numeric input to `week_chart()`.
 
 # clockplot 0.6.3.9000
 
 * Fixed bug while allowing free time format. `clock_chart()` and 
-`clock_chart_qlt` were plotting hands along 24 hours. 
+`clock_chart_qlt()` were plotting hands along 24 hours.
 * User does not have give time in `HH:MM:SS`; `HH:MM` or even `H:M` are also
 allowed. 
 * Fixed a bug containing `hour`, `minute`, and `second` variable