+++ NEWS.md
@@ -1,7 +1,11 @@
 # clockplot (development version)
 
-* Added a few tests to `clock_chart_qnt` as well as `clock_chart_len` and `clock_chart_col`
-* Removed bug from `week_chart()` in `stopifnot` part
+* Documented and described all the functions.
+
+# clockplot 0.6.5.9000
+
+* Added a few tests to `clock_chart_qnt` as well as `clock_chart_len` and `clock_chart_col`.
+* Removed bug from `week_chart()` in `stopifnot` part.
 * Added US Accidents data and example. 
 * Added error for non-numeric input to `week_chart()`.